+++ mod-inventory-storage-server/src/main/resources/templates/db_scripts/instance/createInstanceCatalogedDateIndex.sql
@@ -1,0 +1,2 @@
+CREATE INDEX IF NOT EXISTS instance_cataloged_date_idx ON ${myuniversity}_${mymodule}.instance
+  (${myuniversity}_${mymodule}.to_timestamp(jsonb->>'catalogedDate'));