+++ mod-inventory-storage-server/src/main/resources/templates/db_scripts/instance/createInstanceCreationDateIndex.sql
@@ -1,0 +1,2 @@
+CREATE INDEX IF NOT EXISTS instance_creation_date_idx ON ${myuniversity}_${mymodule}.instance
+  (creation_date);